requirements.txt
# Ingestion dependencies
openai>=1.30.0
pymongo>=4.7.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0

# Backend dependencies
fastapi>=0.111.0
uvicorn>=0.30.1
pandas>=2.2.0
python-docx>=1.1.2
openpyxl>=3.1.2
python-multipart>=0.0.9
